refactor: extract ContainerPlugin abstract base class
Move shared Docker logic (imageBuild resolution, env resolution,
isWorkflowAgentContext, normalizeRepoUrls) into ContainerPlugin.
Both BaseContainerAgentPlugin and ScriptContainerPlugin now extend it,
eliminating duplicated code.
